// src/enums/AppColors.ts
enum AppColors {
  // Helper colors.
  TRANSPARENT = 'transparent',

  // Theme colors.
  // TODO: Edit theme colors to match app theme.
  // Base seed color for the theme.
  SEED = '#6750A4',
  // Light theme colors.
  THEME_LIGHT_PRIMARY = '#6750A4',
  THEME_LIGHT_ON_PRIMARY = '#FFFFFF',
  THEME_LIGHT_PRIMARY_CONTAINER = '#EADDFF',
  THEME_LIGHT_ON_PRIMARY_CONTAINER = '#21005D',
  THEME_LIGHT_SECONDARY = '#625B71',
  THEME_LIGHT_ON_SECONDARY = '#FFFFFF',
  THEME_LIGHT_SECONDARY_CONTAINER = '#E8DEF8',
  THEME_LIGHT_ON_SECONDARY_CONTAINER = '#1D192B',
  THEME_LIGHT_TERTIARY = '#7D5260',
  THEME_LIGHT_ON_TERTIARY = '#FFFFFF',
  THEME_LIGHT_TERTIARY_CONTAINER = '#FFD8E4',
  THEME_LIGHT_ON_TERTIARY_CONTAINER = '#31111D',
  THEME_LIGHT_ERROR = '#B3261E',
  THEME_LIGHT_ON_ERROR = '#FFFFFF',
  THEME_LIGHT_ERROR_CONTAINER = '#F9DEDC',
  THEME_LIGHT_ON_ERROR_CONTAINER = '#410E0B',
  THEME_LIGHT_BACKGROUND = '#FFFBFE',
  THEME_LIGHT_ON_BACKGROUND = '#1C1B1F',
  THEME_LIGHT_SURFACE = '#FFFBFE',
  THEME_LIGHT_ON_SURFACE = '#1C1B1F',
  THEME_LIGHT_SURFACE_VARIANT = '#E7E0EC',
  THEME_LIGHT_ON_SURFACE_VARIANT = '#49454F',
  THEME_LIGHT_OUTLINE = '#79747E',
  THEME_LIGHT_OUTLINE_VARIANT = '#CAC4D0',
  THEME_LIGHT_SHADOW = '#000000',
  THEME_LIGHT_SCRIM = '#000000',
  THEME_LIGHT_INVERSE_SURFACE = '#313033',
  THEME_LIGHT_INVERSE_ON_SURFACE = '#F4EFF4',
  THEME_LIGHT_INVERSE_PRIMARY = '#D0BCFF',
  THEME_LIGHT_ELEVATION_LEVEL_0 = TRANSPARENT,
  THEME_LIGHT_ELEVATION_LEVEL_1 = 'rgb(247, 242, 250)',
  THEME_LIGHT_ELEVATION_LEVEL_2 = 'rgb(243, 237, 248)',
  THEME_LIGHT_ELEVATION_LEVEL_3 = 'rgb(238, 232, 245)',
  THEME_LIGHT_ELEVATION_LEVEL_4 = 'rgb(237, 231, 244)',
  THEME_LIGHT_ELEVATION_LEVEL_5 = 'rgb(234, 227, 242)',
  THEME_LIGHT_SURFACE_DISABLED = 'rgba(28, 27, 31, 0.12)',
  THEME_LIGHT_ON_SURFACE_DISABLED = 'rgba(28, 27, 31, 0.38)',
  THEME_LIGHT_BACKDROP = 'rgba(50, 47, 56, 0.4)',
  // Dark theme colors.
  THEME_DARK_PRIMARY = '#D0BCFF',
  THEME_DARK_ON_PRIMARY = '#381E72',
  THEME_DARK_PRIMARY_CONTAINER = '#4F378B',
  THEME_DARK_ON_PRIMARY_CONTAINER = '#EADDFF',
  THEME_DARK_SECONDARY = '#CCC2DC',
  THEME_DARK_ON_SECONDARY = '#332D41',
  THEME_DARK_SECONDARY_CONTAINER = '#4A4458',
  THEME_DARK_ON_SECONDARY_CONTAINER = '#E8DEF8',
  THEME_DARK_TERTIARY = '#EFB8C8',
  THEME_DARK_ON_TERTIARY = '#492532',
  THEME_DARK_TERTIARY_CONTAINER = '#633B48',
  THEME_DARK_ON_TERTIARY_CONTAINER = '#FFD8E4',
  THEME_DARK_ERROR = '#F2B8B5',
  THEME_DARK_ON_ERROR = '#601410',
  THEME_DARK_ERROR_CONTAINER = '#8C1D18',
  THEME_DARK_ON_ERROR_CONTAINER = '#F9DEDC',
  THEME_DARK_BACKGROUND = '#1C1B1F',
  THEME_DARK_ON_BACKGROUND = '#E6E1E5',
  THEME_DARK_SURFACE = '#1C1B1F',
  THEME_DARK_ON_SURFACE = '#E6E1E5',
  THEME_DARK_SURFACE_VARIANT = '#49454F',
  THEME_DARK_ON_SURFACE_VARIANT = '#CAC4D0',
  THEME_DARK_OUTLINE = '#938F99',
  THEME_DARK_OUTLINE_VARIANT = '#49454F',
  THEME_DARK_SHADOW = '#000000',
  THEME_DARK_SCRIM = '#000000',
  THEME_DARK_INVERSE_SURFACE = '#E6E1E5',
  THEME_DARK_INVERSE_ON_SURFACE = '#313033',
  THEME_DARK_INVERSE_PRIMARY = '#6750A4',
  THEME_DARK_ELEVATION_LEVEL_0 = TRANSPARENT,
  THEME_DARK_ELEVATION_LEVEL_1 = 'rgb(37, 35, 41)',
  THEME_DARK_ELEVATION_LEVEL_2 = 'rgb(42, 40, 48)',
  THEME_DARK_ELEVATION_LEVEL_3 = 'rgb(48, 45, 55)',
  THEME_DARK_ELEVATION_LEVEL_4 = 'rgb(50, 46, 57)',
  THEME_DARK_ELEVATION_LEVEL_5 = 'rgb(53, 50, 62)',
  THEME_DARK_SURFACE_DISABLED = 'rgba(230, 225, 229, 0.12)',
  THEME_DARK_ON_SURFACE_DISABLED = 'rgba(230, 225, 229, 0.38)',
  THEME_DARK_BACKDROP = 'rgba(50, 47, 56, 0.4)',

  // App colors.
  // TODO: Add app colors here.
}

export default AppColors;